Fixed failing test
