+++ packages/api/Tests/UnitTests/AccessToken/AccessTokenStorageTests.cs
@@ -91,56 +91,5 @@ public class AccessTokenStorageTests
         var tokens = await storage.GetActiveTokens(user);
         Assert.Empty(tokens);
     }
-
-    [Fact]
-    public async Task Authenticate_ExpiredUser_WithValidPAT_ShouldFail()
-    {
-        // Arrange
-        var dbContext = new AlvTimeDbContextBuilder()
-            .WithPersonalAccessTokens()
-            .WithUsers()
-            .CreateDbContext();
-
-        var storage = new UserRepository(dbContext);
-
-        // This user and the following access token are provided by the WithUsers|WithPATs methods above
-        await storage.UpdateUser(new UserDto
-        {
-            Id = 1,
-            EndDate = DateTime.Now.AddMonths(-1)
-        });
-
-        var accessToken = await dbContext.AccessTokens.FindAsync(1);
-
-        var httpContext = new DefaultHttpContext();
-        httpContext.Request.Headers.Add(HeaderNames.Authorization, $"Bearer {accessToken?.Value}");
-
-        // The options and loggerFactory objects must return these values. If not, we get a NullPtrException
-        // Source: https://stackoverflow.com/questions/58963133/unit-test-custom-authenticationhandler-middleware
-        var options = new Mock<IOptionsMonitor<PersonalAccessTokenOptions>>();
-        options.Setup(x => x.Get(It.IsAny<string>()))
-            .Returns(new PersonalAccessTokenOptions());
-
-        var logger = new Mock<ILogger<PersonalAccessTokenHandler>>();
-        var loggerFactory = new Mock<ILoggerFactory>();
-        loggerFactory.Setup(x => x.CreateLogger(It.IsAny<string>())).Returns(logger.Object);
-
-        var authenticationHandler = new PersonalAccessTokenHandler(
-            options.Object,
-            loggerFactory.Object,
-            UrlEncoder.Default,
-            storage,
-            Mock.Of<SystemClock>());
-
-        await authenticationHandler.InitializeAsync(
-            new AuthenticationScheme(nameof(PersonalAccessTokenHandler), null, typeof(PersonalAccessTokenHandler)),
-            httpContext);
-
-        // Act
-        var result = await authenticationHandler.AuthenticateAsync();
-
-        // Assert
-        Assert.False(result.Succeeded);
-        Assert.Equal("The user has an end date in the past", result.Failure?.Message);
-    }
+    
 }
